next_gen/pyproject.toml
[project]
name = "next-gen"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.14.2",
    "lxml>=6.0.2",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
    "requests>=2.32.5",
    "yfinance>=0.2.66",
]
